Add stub engine for Dawn native
Basic methods always succeed, even though most of them do nothing.

It does create a Dawn device, proving that we can link to the Dawn
native library.

Fix the method for finding Dawn.  The dawn_native library has a link
dependency to the dawn (core) library.  So we have to find and configure
that too.