Process command before getting FB info
Process the graphics commands to ensure the framebuffer buffer
is mapped into host memory before getting the framebuffer info.
That's needed to ensure the buffer's CPU pointer is valid.

This is necessary for the Dawn flow, which doesn't promise that
the host-side pointer is the same every time you map the buffer.