Extend engine configuration options.
Currently an InitializeWithDevice method is provided to the engines
which is populated with the device given when calling Execute. This CL
extends that mechanism to have EngineConfig structs for each engine.
Those structs are given to Execute and passed to the engine. This will
allow extending the Vulkan code to accept PhysicalDevice and available
feature/extension information along with the device.