Flip the sample app to be less quiet.
The sample app outputting just an error code by default is confusing to
users. The -s flag was originally created to turn on the summary mode.
This CL removes -s from the help (but it still parses) and adds a -q
flag instead. The quiet flag is used by the test runner to suppress the
summary output.