[vulkan] Generate wrapper functions for vulkan methods. (#328)
This CL adds a update_vk_wrappers.py script which is executed during the
build. The script will generate src/vk-wrappers.h and
src/vk-wrappers.inc files into the binary directory. The .h file lists
the struct entries for the VulkanPtr struct.

The vk-wrappers.inc file contains the code to load the vulkan function
pointers. If the vk.xml file is found, the wrappers will be lambdas, if
vk.xml is not found they're straight function pointers. This should
allow us to work with the CTS as we do now.

Issue #324.