Allow multiple pipelines to execute.
This CL updates amber (and the vulkan backend) to work with multiple
pipelines. The vulkan backend is changed to store a PipelineData
structure with the colour/depth formats, vertex data and the vulkan
pipeline for a given amber::Pipeline.

Each command is changed to take the pipeline it is associated with,
those pipelines are used when executing the commands.