Force cmake settings into the cache.
If you have settings which are defined by OPTIONS in the subprojects, you
can not set those options if the cmake version is too low. There is a
policy CMP0077 which can be set, but that gets reset as soon as the
cmake_minimium_required is called (which happens in every subproject).

This CL forces all of the things we care about into the cmake cache so
they don't get reset by the downstream project. This only seems to be
an issue on the Windows build but removes ~25 targets from a build
as we correctly turn off tests and things.