[vulkan] simplify image transition code. (#441)
This CL splits the ChangeLayout method into a method to change to Draw
mode and a method to change to Probe mode. The framebuffer is always
transitioned to Draw mode after initialization. The RenderPassGuard now
transitions the framebuffer into draw layout on initialize and out of
draw layout on shutdown.