Generate a Format from a DataBuffer.
When sending a VertexBuffer to Vulkan the data format is passed as a
VkFormat string. When the buffer is provided in AmberScript it is not a
FORMAT buffer. This CL adds the necessary code to generate a Format from
a DatumType. This will allow creating the necessary format to send to
Vulkan.