Unify buffer sizes.
Currently the format buffer and data buffer count their size
differently. For data buffers we counted elements (so 1 vec4 was size 1)
and format buffers counted values (so 1 vec4 was size 4).

This CL unifies the two sizes and adds ElementCount and ValueCount to
make it explicit which version the user wants.

As part of this, the data types used for ssbo and uniform commands must
be consistent. The tests where we switched data formats part way through
are updated.