Only resize buffer if it will become bigger.
When setting data into the buffer we were resizing the buffer to fit the
values provided. This is incorrect as a buffer may have been sized
explicitly with a size command and we don't want to make it smaller.

This CL changes the code to only resize the buffer bigger and not
smaller.

Fixes #503