[vkscript] Set default format into buffers.
When the SSBO size command is used there is currently no format set into
a buffer. This CL sets a default format of `R8_SINT`. A flag is also set
that the format is defaulted and allows it to be overridden.

This permits the usage of scratch buffers, but allows the buffer to be
transferred to/from the device as we can set the buffer size to the
number of provided elements.

Fixes #513
Issue #501