Use simple include paths (#571)
This change allows DXC to be picked up from any location, which may
not be in third_party/dxc/include/. A different build system or
environment may place it in a different directory. If amber is part
of another project that already has a copy of DXC, we would not
want two of them.