DEPS: Downgrade glslang to match Dawn
When using Amber with Dawn, both Dawn and Amber pull in Glslang.
If the versions of Glslang differ in ABIs to which Shaderc is
sensitive, then we can't use Dawn with Amber.  The symptom is
that Amber won't link.

In this case:
Glslang commit 137d2ba changed the signature of TShader::mapIO,
and that was merged into Glslang master with f04f1f93 (after
an earlier commit and revert pair).

So downgrade the glslang commit to the one currently supported
in Dawn.