[AmberScript] Fix parsing of matrix data into buffers.
Currently parsing of matrix data in a buffer DATA block would trigger
memory corrpution issues. This was due to several things:

 * Incorrect interpretation of mat2x4 for rows and columns
 * Not clamping the amount of data read from the DATA block
 * Handling of padding values.

With this CL we correctly handle mat2x4 as 2 columns and 4 rows. We do
not require padding values in the DATA block for matrices, it will be
inserted automatically as needed. We error out if the number of DATA
elements is not a multiple of the amount of input needed for the number
of elements.

Fixes #616