Remove GetComponents from the Format class. (#674)
With this change, it is no longer possible to get a list of components
from a Format. Instead, the segments should be used and the padding
taken into account.

A GetOnlyComponent was added for the usage in the amberscript parser but
requires there to only be a single component in the format.