Remove buffer type from BufferInfo struct.
The buffer type can get changed late in the process (this happens in the
case of OpenCL where we determine the buffer type from binding data that
is seen later). This means, the buffer type can be changed _after_ the
BufferInfo structure is setup.

Make sure we always retrieve the buffer type from the buffer directly
instead of caching.