Add ability to extract shader source.
If the shader is compiled by Amber, the compiled shader data will now be
returned when querying for shader information. This allows adding a flag
'-w <filename>' to the amber sample program to dump the SPIR-V
disassembly of the compiled shaders to a file.

This also fixes up the shader name and optimization pass return values
in the shader_info to hold the correct information based on pipeline.