Fix -V driver version info
We now request "VK_KHR_get_physical_device_properties2" when creating the Vulkan instance, if it is available and if "-V" is passed. We also now correctly get the address of the "vkGetPhysicalDeviceProperties2KHR" function (*with* the KHR suffix), as opposed to "vkGetPhysicalDeviceProperties2", which is only available in Vulkan 1.1.

Fix #735