Debugger: Add amber::debug namespace
This namespace holds the types used to describe the tests to perform on a shader debugger.
`debug.cc` implements the `amber::debug::Script` class, which is a simple record-and-replay implementation of the debugger interfaces. This is used to hold the parsed debugger instructions, and can be replayed onto the real debugger when the commands are executed.

The parser for generating the debug scripts along with the debugger itself come in later changes.