Debugger: Don't stop progressing on error
The `Step`XXX`()` functions and `Continue()` would be a no-op if any errors have been previously raised for the thread.
In retrospect this isn't very helpful - it just causes the `Expect`XXX`()` functions to produce nonsense repeated failures.
Instead, keep on going, error or not.

Also include the line content when we encounter a line mismatch.