Implement draw_grid
Add support for draw_grid, which is fairly similar to draw_rect,
except it also takes parameters to split the rectangle into
desired number of columns and rows of cells.

This functionality is meant for the testing of vertex shaders,
giving an easy way to produce a lot of vertices.

The implementation is based on draw_rect. Dawn rendering is not
implemented.