Add support for 8-bit int shader/storage features (#792)
This pull request adds support for 8-bit integers to be used in shaders, both as a basic type and for input/output as part of push constants and uniform/storage blocks. It adds 4 new features understood by Amberscript which are needed to request the corresponding Vulkan features. The original structure is called "8BitStorageFeatures" but it has been renamed to "Storage8BitFeatures" due to the leading digit causing problems to the parser.