Make topologies upper-case in documentation, and fix count in example
Amber requires topologies to be upper-case.  Furthermore, when
processing a triangle list in a DRAW_ARRAY command, Amber expects
COUNT to specify the number of triangles, not the number of vertices
overall.  The documentation and examples featured lower-case
topologies and a case where a triangle list's count was being derived
from the number of vertices; this change fixes these issues.

Fixes #818.
Fixes #819.