Debugger: Scan for VIRTUAL_FILEs and wait for steps
If the `dap::Source` does not provide the embedded source, scan the virtual file system for the file by path.

Also fix a race condition where we would instruct the debugger to step, but not actually wait for it to notify us that the step had completed before making further checks.

Updated the `debugger_hlsl_line_stepping.amber` to:
* Use `VIRTUAL_FILE` for the file content
* Fix the scoping of `CompilationUnit` <- `Function` <- `LexicalBlock`
* Remove the unused and incorrect debug info from the fragment shader. We'll have proper testing for this once DXC produces correct debug info.