Fix Python warnings related to 'is' operator (#891)
Replace the 'is' operators with '==' when there is a literal among the
arguments. The Python 'is' operator is not meant to be used with
literals, and the code emitted warnings like:

tools/git-sync-deps:231: SyntaxWarning: "is" with a literal. Did you mean "=="?

For the origin of this warning in Python, see
https://bugs.python.org/issue34850

Fix git-sync-deps flags in CI scripts.

Fix use of global variables.

Fix: #890