Debugger: Don't use hardcoded debugger port
SwiftShader uses the `VK_DEBUGGER_PORT` to determine the port for the DAP server. Use the same method in amber so they're always in sync.

Ideally we wouldn't be using environment variables, nor sockets for two endpoints in the same process, but we haven't got a C API defined yet.