Fix Vulkan synchronization validation errors
Vulkan CTS TransferImage::ImageBarrier() is often used to
synchronize images which are then used as attachments in
a RenderPass. Depending on the load operation used, the RenderPass
may need either read or write access. (see Section 8.1 Render Pass
Creation in the Vulkan 1.2.x specification).

To allow any load operation to work correctly, set both read and write
access in the dstAccessMask when transitioning to an attachment layout.