Remove shader debugging from Amber
SwiftShader was the only implementation that supported this, but has
disabled this code due to lack of use. If there's no driver that
supports shader debugging, then it doesn't need to exist and be
maintained in Amber.

If one day, we want this again, then this CL can be reverted.

Fixes: #999