Modify Android test files (#1021)
To get android build for amber to work internally, I need a few changes. I do
not know enough about the junit testing to know if this will mess anyonly else
up.

I'm also not sure who else uses all of this. Let me know if this causes a
problem.