Fixup acceleration structure and raytracing flags.
When determining if acceleration structures and ray tracing are
supported it isn't enough to just check for the device extension. Both
of these extensions maybe available, then the features contain a flag
which determines if it is implemented. So, if we've got the features, we
need to additionally check the flag and reset our internal tracking
based on the implementation flag.

Change-Id: I09b04007a47e5b5bec34c7ed910949401ca18a3b