Group members of ConfigHelperVulkan.
Move the various members into separate structures to a) group them and
b) make the names more distinct from other versions used in the helper.

This adds a `vk_` group for all the vulkan objects, a `supports_` struct
for all the booleans and a `features_` struct for all the feature
objects.